1. 1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -15,7 +15,7 @@ void print_binary(unsigned long int n)
 		printf("0");
 		return;
 	}
-	current_bit = 1UL << (sizeof(unsigned long int) * 8 - 1);
+	current_bit = (unsigned long int)1 << (sizeof(unsigned long int) * 8 - 1);
 	leading = 1;
 	while (current_bit > 0)
 	{
